+++ scripts/__init__.py
@@ -3,4 +3,3 @@ from .get_metadata import (get_show_metadata,
                            _format_episode_metadata,
                            get_episode_metadata,
                            write_episode_metadata)
-from .visualization_generator import VisualizationGenerator
