+++ thronetalk-game-of-thrones-summarizer/data/show_metadata.json
