+++ thronetalk-game-of-thrones-summarizer/tests/mock_functions.py
