+++ thronetalk-game-of-thrones-summarizer/utils/data_analysis.py
