+++ thronetalk-game-of-thrones-summarizer/utils/model.py
